components_by_event: array -> struct
components_by_event is now structured like this: array [event_type] -> struct[$ event_number] -> array of components. this is because in some cases (like collision events) the event_number is very high, like 100000, causing a lot of memory to be wasted.

the methods are now static to optimize memory.

fixed remove_component: now it always detaches the component, and gives the option to destroy it or keep it as it is.